+++ main.py
@@ -1,8 +1,9 @@
 import asyncio
 import os
-import warnings
+import typing
 
 import discord
+from discord import app_commands
 from dotenv import load_dotenv
 
 from functions import *
@@ -15,10 +16,14 @@ from utilities import *
 # Implement whitelist
 
 
+# UX variables
+MESSAGE_TIMEOUT = 3
+
 # Load environment
 load_dotenv()
 BOT_TOKEN = os.getenv('BOT_TOKEN')
-HOME_GUILD = os.getenv('HOME_GUILD')
+HOME_GUILD_ID = int(os.getenv('HOME_GUILD_ID'))
+DEV_ID = int(os.getenv('DEV_ID'))
 
 # Initialize bot
 intents = discord.Intents().none()
@@ -28,19 +33,21 @@ intents.messages = True
 intents.message_content = True
 intents.voice_states = True
 
+# Create client
 client = discord.Client(command_prefix='!', intents=intents)
+tree = app_commands.CommandTree(client)
+
+# Get sync guild
+sync_guild = discord.Object(id=HOME_GUILD_ID)
 
 
 # Initialization behavior
 @client.event
 async def on_ready():
-    # Check for home guild (verbose output)
-    # for guild in client.guilds:
-    #     if int(guild.id) == int(HOME_GUILD):
-    #         print('Bot connected to debug server.')
-    #         break
-    # else:
-    #     warnings.warn('Bot not connected to verbose server.')
+    # Sync commands to home guild
+    guild = client.get_guild(int(HOME_GUILD_ID))
+    # await tree.sync(guild=guild)
+    print(f'Commands synced with {guild.name} ({guild.id})')
 
     # Status
     await client.change_presence(status=discord.Status.online, activity=discord.Game('🎵some tunes🎵'))
@@ -55,63 +62,82 @@ async def on_ready():
 
 
 # Commands
-@client.event
-async def on_message(message):
-    # Parse
-    command = message.content.split(' ')[0]
-    if not command[0] == '!': return
-    command = command[1:].lower()
-
-    # Get loop
+@tree.command(name='sync', description='Synchronize commands globally', guild=discord.Object(id=HOME_GUILD_ID))
+async def command_join(interaction):
+    await interaction.response.defer(ephemeral=True)
+    # Reject if not dev
+    if interaction.user.id != DEV_ID:
+        await interaction.followup.send('❌')
+        return
+    # Sync all commands
     loop = asyncio.get_running_loop()
+    loop.create_task(tree.sync())
+    await interaction.followup.send('👍')
 
-    # Connect to voice channel
-    if command == 'join':
-        await message.add_reaction('👍')
-        loop.create_task(type_during(message, join(client, message)))
 
-    # Disconnect
-    elif command == 'leave':
-        await message.add_reaction('👍')
-        loop.create_task(type_during(message, disconnect(client, message)))
+@tree.command(name='join', description='Join the voice channel', guild=sync_guild)
+async def command_join(interaction):
+    await interaction.response.defer(ephemeral=True)
+    loop = asyncio.get_running_loop()
+    loop.create_task(join(client, interaction))
+    await interaction.followup.send('👍')
 
-    # Play music
-    elif command == 'play':
-        await message.add_reaction('👍')
-        loop.create_task(type_during(message, play(client, message)))
 
-    # Skip current song
-    elif command == 'skip':
-        await message.add_reaction('👍')
-        loop.create_task(type_during(message, skip(client, message)))
+@tree.command(name='leave', description='Leave the voice channel', guild=sync_guild)
+async def command_leave(interaction):
+    await interaction.response.defer(ephemeral=True)
+    loop = asyncio.get_running_loop()
+    loop.create_task(leave(client, interaction))
 
-    # Pause current song
-    elif command == 'pause':
-        await message.add_reaction('👍')
-        loop.create_task(type_during(message, pause(client, message)))
 
-    # Resume current song
-    elif command == 'resume':
-        await message.add_reaction('👍')
-        loop.create_task(type_during(message, resume(client, message)))
+@tree.command(name='play', description='Play a song', guild=sync_guild)
+async def command_play(interaction, url: str):
+    await interaction.response.defer(ephemeral=True)
+    loop = asyncio.get_running_loop()
+    loop.create_task(play(client, interaction, url=url))
 
-    # Preview current song
-    elif command == 'playing' or command == 'current':
-        await message.add_reaction('👍')
-        loop.create_task(type_during(message, previewCurrent(client, message)))
 
-    # Preview next song
-    elif command == 'next':
-        await message.add_reaction('👍')
-        loop.create_task(type_during(message, previewNext(client, message)))
+@tree.command(name='skip', description='Skip a song', guild=sync_guild)
+async def command_skip(interaction, index: typing.Optional[int]):
+    await interaction.response.defer(ephemeral=True)
+    loop = asyncio.get_running_loop()
+    loop.create_task(skip(client, interaction, index=index))
 
-    # Preview queue
-    elif command == 'queue':
-        await message.add_reaction('👍')
-        loop.create_task(type_during(message, previewQueue(client, message)))
 
-    else:
-        await message.add_reaction('❌')
+@tree.command(name='pause', description='Pause the current song', guild=sync_guild)
+async def command_pause(interaction):
+    await interaction.response.defer(ephemeral=True)
+    loop = asyncio.get_running_loop()
+    loop.create_task(pause(client, interaction))
+
+
+@tree.command(name='resume', description='Resume the current song', guild=sync_guild)
+async def command_resume(interaction):
+    await interaction.response.defer(ephemeral=True)
+    loop = asyncio.get_running_loop()
+    loop.create_task(resume(client, interaction))
+
+
+@tree.command(name='current', description='Show the currently playing song', guild=sync_guild)
+async def command_playing(interaction):
+    await interaction.response.defer(ephemeral=True)
+    loop = asyncio.get_running_loop()
+    loop.create_task(previewCurrent(client, interaction))
+
+
+@tree.command(name='next', description='Show the next song', guild=sync_guild)
+async def command_next(interaction):
+    await interaction.response.defer(ephemeral=True)
+    loop = asyncio.get_running_loop()
+    loop.create_task(previewNext(client, interaction))
+
+
+@tree.command(name='queue', description='Show the song queue', guild=sync_guild)
+async def command_queue(interaction):
+    await interaction.response.defer(ephemeral=True)
+    loop = asyncio.get_running_loop()
+    loop.create_task(previewQueue(client, interaction))
+
 
 # Run
 client.run(BOT_TOKEN)